wip: reorganize http adapters
