+++ src/kairos/workspace/adapter/http/response/create-workspace-response.ts
