+++ src/kairos/workspace/application/create-worskpace.compose.ts
@@ -1,7 +1,7 @@
 import { workspacesCollection } from '@/framework/mongodb/collection/workspaces-collection';
+import { makeInsertWorkspace } from '@/kairos/workspace/adapter/repository/insert-workspace';
 import { makeCreateWorkspace } from '@/kairos/workspace/application/create-workspace';
 import type { InsertWorkspace } from '@/kairos/workspace/application/insert-workspace.type';
-import { makeInsertWorkspace } from '@/kairos/workspace/infrastructure/repository/insert-workspace';
 
 const insertWorkspace: InsertWorkspace = makeInsertWorkspace(workspacesCollection);
